Move DefaultCodeGeneratorService to the generators folder.
