+++ src/main/java/com/rocketden/main/service/generators/JavaDefaultCodeGeneratorService.java
@@ -7,7 +7,6 @@ import com.rocketden.main.exception.api.ApiException;
 import com.rocketden.main.game_object.CodeLanguage;
 import com.rocketden.main.model.problem.ProblemIOType;
 import com.rocketden.main.model.problem.ProblemInput;
-import com.rocketden.main.service.DefaultCodeGeneratorService;
 
 import org.springframework.stereotype.Service;
