+++ src/main/java/com/rocketden/main/service/generators/PythonDefaultCodeGeneratorService.java
@@ -5,7 +5,6 @@ import java.util.List;
 import com.rocketden.main.game_object.CodeLanguage;
 import com.rocketden.main.model.problem.ProblemIOType;
 import com.rocketden.main.model.problem.ProblemInput;
-import com.rocketden.main.service.DefaultCodeGeneratorService;
 
 import org.springframework.stereotype.Service;
